✨ Request 함수 추가
