+++ Bamboo/Resources/Network/Constants.swift
@@ -5,6 +5,7 @@
 //  Created by Mercen on 2023/02/19.
 //
 
+// MARK: - API Urls
 public let baseAPI: String = "http://13.209.33.18:8080"
 public let userAPI: String = "\(baseAPI)/user"
 public let postAPI: String = "\(baseAPI)/post"
